python/lib/pyduino-sensor/pyproject.toml
# https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html
[build-system]
requires = [
    "setuptools >= 40.9.0", # PEP 517
]
build-backend = "setuptools.build_meta"

[project]
name = "pyduino-sensor"
description = "Hydrogeologger/Pyduino sensor hardware and post processing helper package"
authors = [
    { name = "Lieven Tan" }
]
dynamic = ["version"]
requires-python = ">=2.7"
keywords = ["pyduino", "sensor"]
classifiers = [
    # Project maturity
    #   3 - Alpha
    #   4 - Beta
    #   5 - Production/Stable
    "Development Status :: 3 - Alpha",

    # Python version support
    "Programming Language :: Python"
]
dependencies = [
    "gpiozero; platform_machine == 'armv6l'",
    "pyserial; platform_machine == 'armv61'"
]

[project.optional-dependencies]
rpi = [
    "gpiozero",
    "pyserial"
]
dev = [
    "pyduino-sensor[rpi]",
    "pytest"
]

[project.urls]
"Homepage" = "https://github.com/hydrogeologger/pyduino"
"Bug Tracker" = "https://github.com/hydrogeologger/pyduino/issues"

[tool.setuptools.dynamic]
version = { attr = "pyduino_sensor.__version__" }

[tool.setuptools.package-dir] # Directory containing packages
"" = "src"

[tool.setuptools.packages.find]
where = ["src"] # list of folders that contain the packages (["."] by default)
include = [
    "pyduino_sensor*"
] # package names should match these glob patterns (["*"] by default)
exclude = [] # exclude packages matching these glob patterns (empty by default)
